+++ test/interpreter/bbcode.jl
@@ -62,13 +62,13 @@ end
         @test all(map(==, ir.stmts.line, new_ir.stmts.line))
         @test all(map(==, ir.stmts.flag, new_ir.stmts.flag))
         @test length(Mooncake.collect_stmts(bb_code)) == length(stmt(ir.stmts))
-        @test Mooncake.id_to_line_map(bb_code) isa Dict{ID,Int}
+        @test Mooncake.BBCodes.id_to_line_map(bb_code) isa Dict{ID,Int}
     end
     @testset "control_flow_graph" begin
         ir = Base.code_ircode_by_type(Tuple{typeof(sin),Float64})[1][1]
         bb = BBCode(ir)
         new_ir = Core.Compiler.IRCode(bb)
-        cfg = Mooncake.control_flow_graph(bb)
+        cfg = Mooncake.BBCodes.control_flow_graph(bb)
         @test all(map((l, r) -> l.stmts == r.stmts, ir.cfg.blocks, cfg.blocks))
         @test all(map((l, r) -> sort(l.preds) == sort(r.preds), ir.cfg.blocks, cfg.blocks))
         @test all(map((l, r) -> sort(l.succs) == sort(r.succs), ir.cfg.blocks, cfg.blocks))
@@ -152,51 +152,51 @@ end
             @testset "Expr" begin
                 id = ID()
                 d = Dict{ID,Bool}(id => false)
-                Mooncake._find_id_uses!(d, Expr(:call, sin, 5))
+                Mooncake.BBCodes._find_id_uses!(d, Expr(:call, sin, 5))
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, Expr(:call, sin, id))
+                Mooncake.BBCodes._find_id_uses!(d, Expr(:call, sin, id))
                 @test d[id] == true
             end
             @testset "IDGotoIfNot" begin
                 id = ID()
                 d = Dict{ID,Bool}(id => false)
-                Mooncake._find_id_uses!(d, IDGotoIfNot(ID(), ID()))
+                Mooncake.BBCodes._find_id_uses!(d, IDGotoIfNot(ID(), ID()))
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, IDGotoIfNot(true, ID()))
+                Mooncake.BBCodes._find_id_uses!(d, IDGotoIfNot(true, ID()))
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, IDGotoIfNot(id, ID()))
+                Mooncake.BBCodes._find_id_uses!(d, IDGotoIfNot(id, ID()))
                 @test d[id] == true
             end
             @testset "IDGotoNode" begin
                 id = ID()
                 d = Dict{ID,Bool}(id => false)
-                Mooncake._find_id_uses!(d, IDGotoNode(ID()))
+                Mooncake.BBCodes._find_id_uses!(d, IDGotoNode(ID()))
                 @test d[id] == false
             end
             @testset "IDPhiNode" begin
                 id = ID()
                 d = Dict{ID,Bool}(id => false)
-                Mooncake._find_id_uses!(d, IDPhiNode([ID()], Vector{Any}(undef, 1)))
+                Mooncake.BBCodes._find_id_uses!(d, IDPhiNode([ID()], Vector{Any}(undef, 1)))
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, IDPhiNode([ID()], Any[id]))
+                Mooncake.BBCodes._find_id_uses!(d, IDPhiNode([ID()], Any[id]))
                 @test d[id] == true
             end
             @testset "PiNode" begin
                 id = ID()
                 d = Dict{ID,Bool}(id => false)
-                Mooncake._find_id_uses!(d, PiNode(false, Bool))
+                Mooncake.BBCodes._find_id_uses!(d, PiNode(false, Bool))
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, PiNode(id, Bool))
+                Mooncake.BBCodes._find_id_uses!(d, PiNode(id, Bool))
                 @test d[id] == true
             end
             @testset "ReturnNode" begin
                 id = ID()
                 d = Dict{ID,Bool}(id => false)
-                Mooncake._find_id_uses!(d, ReturnNode())
+                Mooncake.BBCodes._find_id_uses!(d, ReturnNode())
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, ReturnNode(5))
+                Mooncake.BBCodes._find_id_uses!(d, ReturnNode(5))
                 @test d[id] == false
-                Mooncake._find_id_uses!(d, ReturnNode(id))
+                Mooncake.BBCodes._find_id_uses!(d, ReturnNode(id))
                 @test d[id] == true
             end
         end
@@ -225,7 +225,7 @@ end
             ],
             Any[Any for _ in 1:4],
         )
-        @test Mooncake._is_reachable(BBCode(ir).blocks) == [true, false, false]
+        @test Mooncake.BBCodes._is_reachable(BBCode(ir).blocks) == [true, false, false]
     end
     @testset "remove_unreachable_blocks!" begin
